+++ src/Network/Network/Models/PSNetworkInterface.cs
@@ -45,6 +45,8 @@ namespace Microsoft.Azure.Commands.Network.Models
 
         public PSNetworkSecurityGroup NetworkSecurityGroup { get; set; }
 
+        public PSResourceId PrivateEndpoint { get; set; }
+
         [Ps1Xml(Target = ViewControl.Table)]
         public string ProvisioningState { get; set; }
 
@@ -76,6 +78,12 @@ namespace Microsoft.Azure.Commands.Network.Models
         public string NetworkSecurityGroupText
         {
             get { return JsonConvert.SerializeObject(NetworkSecurityGroup, Formatting.Indented, new JsonSerializerSettings() { NullValueHandling = NullValueHandling.Ignore }); }
+        }
+
+        [JsonIgnore]
+        public string PrivateEndpointText
+        {
+            get { return JsonConvert.SerializeObject(PrivateEndpoint, Formatting.Indented, new JsonSerializerSettings() { NullValueHandling = NullValueHandling.Ignore }); }
         }
 
         public bool ShouldSerializeIpConfigurations()